server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.37.0
psycopg2-binary
sqlalchemy
alembic
python-dotenv
itsdangerous
python-jose 
passlib[bcrypt]
bcrypt<4.0.0
pydantic[email]==2.12.3
requests
pydantic-settings>=2.0,<3
fastapi-mail==1.4.2
jinja2
python-multipart
pytest==8.4.2
pytest-cov
testcontainers[postgres]
httpx==0.28.1
google-generativeai
json-repair
google-cloud-aiplatform
google-genai
gspread==6.1.4
google-auth==2.36.0
psycopg2-binary==2.9.11

pandas
xgboost==2.0.3
